Introduce function to get stops for a route
This function takes a Route, which has been augmented to include its
string `id` and uses that to request the stops for that route. As
mentioned in the comment, we are not able to request all stops with
route information per:

https://api-v3.mbta.com/docs/swagger/index.html#/Stop/ApiWeb_StopController_index

So we must make a request per-route, but given that we are only dealing
with 8 routes, this shouldn't cause any rate-limiting concerns.

We will use this to resolve question two next.