Eagerly return the best path if we see it
This will keep us from needlessly exploring the branches as a kind of
one-layer peek-ahead to see if the match will happen in the next
recursive call. This allows us to return the shortest path, and should
also generally improve performance, even though we have more iterations
to perform, we should generally avoid searching branches that we do not
need to.