Function to find path of routes to connect to stops
This calls into the function from the previous commit after looking up
all live data from the MBTA API and seeding with empty sets and lists
for the existing route information. It also then prints out the answer
to standard out for the user.

It currently just has two hard-coded values in the main; this will be
replaced with input from the user in the next commit.