Include tests for `collect_stop_data`
This includes heavy happy-path testing for the min/max/stop-route-map
computations, as well as sad-path testing for the route and stop
lookups.

Again, we are skipping tests for `print_stop_data` as it is just the
interface with standard out and is being treated as lower priority. If
we feel we need this tested, we can revisit.

We will include tests for its helper `build_route_list_name` though;
that will be in the next commit.